security added

+++ ex2/ft_plant_growth.py
@@ -1,5 +1,5 @@
 class Plant:
-    def __init__(self, name:str, height:int, age_in_days:int, 
+    def __init__(self, name:str, height:float, age_in_days:int, 
 growth_speed=0.4, max_days=20):
         self.name = name
         self.height = height
